Restore react-router error handling (#5682)
* fix(editor) Restore react-router error handling

* fix(tests) Fix cyclic dependency introduced to tests

* chore(editor) Remove unused export
